Move to core mod for hotbar expansion
Previous, the swapping was handled all client side with some server side logic added in 3.2. This commit moves to a core mod expanding the player inventory to 36 slots so items can be set via "currentItem" instead of relying on the faux-elongated hotbar. The downside is vanilla server support is dropped. A reimplementation is needed that is used as a fallback when the client connects to a vanilla server or a server that does not have the core mod.